Melhora prompt

+++ backend/app/models/transaction.py
@@ -1,5 +1,6 @@
 import uuid
 from datetime import datetime
+from decimal import Decimal
 
 from sqlalchemy import DateTime, ForeignKey, Numeric, String, func
 from sqlalchemy.dialects.postgresql import UUID
@@ -19,7 +20,9 @@ class Transaction(Base):
         UUID(as_uuid=True), ForeignKey("users.id", ondelete="CASCADE"), nullable=False, index=True
     )
     description: Mapped[str] = mapped_column(String(255), nullable=False)
-    amount: Mapped[float] = mapped_column(Numeric(12, 2), nullable=False)
+    # Dinheiro é Decimal, nunca float (seção 4.4 do plano de arquitetura): o
+    # Numeric do Postgres já chega como Decimal e é assim que ele deve circular.
+    amount: Mapped[Decimal] = mapped_column(Numeric(12, 2), nullable=False)
     category: Mapped[str | None] = mapped_column(String(100), nullable=True)
     type: Mapped[str] = mapped_column(String(20), nullable=False)  # ex.: "income" | "expense"
     # ex.: "pix" | "dinheiro" | "débito" | "crédito" | "boleto" | "transferência"
